requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
requests
feedparser
beautifulsoup4
python-dotenv
marshmallow
apscheduler
pygithub
